--- src/PaginatedCollection.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace JPI\Utils;
-
-/**
- * Collection that is from a paginated result.
- */
-class PaginatedCollection extends Collection {
-
-    use PaginatedCollectionTrait;
-}
